+++ packages/core/src/default-settings.js
@@ -1,0 +1,16 @@
+export const DEFAULT_SETTINGS = {
+  animated: false,
+  persistRandomOnExport: true,
+  optimize: true,
+  hideFeedback: false,
+  hideNavigation: false,
+  hidePresets: false,
+  hideScaleToFit: false,
+  initialScaleToFit: true,
+  hideAutoRefresh: false,
+  initialAutoRefresh: true,
+  hideGenerate: false,
+  showMultipleExports: false,
+  ignoreStyles: false,
+  frameRate: 60
+};