+++ packages/create-mechanic/function-templates/canvas-video/function/index.js
@@ -1,18 +1,21 @@
-export const handler = async ({ inputs, mechanic }) => {
+export const handler = async ({ inputs, frame, done, drawLoop, getCanvas }) => {
   const { width, height, text, color1, color2, radiusPercentage, turns } =
     inputs;
 
   const center = [width / 2, height / 2];
   const radius = ((height / 2) * radiusPercentage) / 100;
-  let angle = 0;
 
-  const canvas = document.createElement("canvas");
-  canvas.width = width;
-  canvas.height = height;
+  const { ctx } = getCanvas(width, height);
 
-  const ctx = canvas.getContext("2d");
+  // frameCount has the number of the current frame, this is based on the framerate
+  // your animation is running it. For 60 fps the frame at 1 seconds will be 60, while
+  // at 24 fps it will be 24.
+  //
+  // timestamp has the frame offset in seconds and is always the same, no matter the
+  // framerate.
+  drawLoop(({ frameCount, timestamp }) => {
+    const angle = Math.PI * timestamp;
 
-  const drawFrame = () => {
     ctx.fillStyle = "#F4F4F4";
     ctx.fillRect(0, 0, width, height);
 
@@ -36,19 +39,17 @@ export const handler = async ({ inputs, mechanic }) => {
     ctx.fillStyle = "#000000";
     ctx.font = `${height / 10}px sans-serif`;
     ctx.textAlign = "center";
-    ctx.strokeText(text, width / 2, height - height / 20);
-    ctx.fillText(text, width / 2, height - height / 20);
+
+    const textWithFrameCount = `${text} ${frameCount}`;
+    ctx.strokeText(textWithFrameCount, width / 2, height - height / 20);
+    ctx.fillText(textWithFrameCount, width / 2, height - height / 20);
 
     if (angle < turns * 2 * Math.PI) {
-      mechanic.frame(canvas);
-      angle += (2 * Math.PI) / 100;
-      window.requestAnimationFrame(drawFrame);
+      frame();
     } else {
-      mechanic.done(canvas);
+      done();
     }
-  };
-
-  drawFrame();
+  });
 };
 
 export const inputs = {